added export + force prompt
